- Mesh improvements - Add generateBBOX - Add missing BBOX to generated meshes
