+++ render/include/rawrbox/render/model/instanced.hpp
@@ -124,6 +124,7 @@ namespace rawrbox {
 
 			bgfx::setState(flags, 0);
 			this->_material->postProcess();
+			bgfx::discard();
 		}
 	};
 } // namespace rawrbox
