+++ render/include/rawrbox/render/model/material/lit.hpp
@@ -34,19 +34,19 @@ namespace rawrbox {
 		void process(const rawrbox::Mesh<T>& mesh) {
 			rawrbox::MaterialBase::process(mesh);
 
-			if (mesh.specularTexture != nullptr && mesh.specularTexture->valid() && !mesh.wireframe) {
+			if (mesh.specularTexture != nullptr && mesh.specularTexture->valid() && !mesh.lineMode && !mesh.wireframe) {
 				bgfx::setTexture(2, s_texSpecularColor, mesh.specularTexture->getHandle());
 			} else {
 				bgfx::setTexture(2, s_texSpecularColor, rawrbox::BLACK_TEXTURE->getHandle());
 			}
 
-			if (mesh.emissionTexture != nullptr && mesh.emissionTexture->valid() && !mesh.wireframe) {
+			if (mesh.emissionTexture != nullptr && mesh.emissionTexture->valid() && !mesh.lineMode && !mesh.wireframe) {
 				bgfx::setTexture(3, s_texEmissionColor, mesh.emissionTexture->getHandle());
 			} else {
 				bgfx::setTexture(3, s_texEmissionColor, rawrbox::BLACK_TEXTURE->getHandle());
 			}
 
-			if (mesh.opacityTexture != nullptr && mesh.opacityTexture->valid() && !mesh.wireframe) {
+			if (mesh.opacityTexture != nullptr && mesh.opacityTexture->valid() && !mesh.lineMode && !mesh.wireframe) {
 				bgfx::setTexture(4, s_texOpacityColor, mesh.opacityTexture->getHandle());
 			} else {
 				bgfx::setTexture(4, s_texOpacityColor, rawrbox::WHITE_TEXTURE->getHandle());
