+++ render/include/rawrbox/render/model/sprite.hpp
@@ -59,6 +59,7 @@ namespace rawrbox {
 
 				bgfx::setState(flags, 0);
 				this->_material->postProcess();
+				bgfx::discard();
 			}
 		}
 	};
