+++ render/shaders/model/skinned/unlit/vs_model_skinned_unlit.sc
@@ -4,10 +4,10 @@ $output v_color0, v_texcoord0
 #include <bgfx_shader.sh>
 
 #include <../../../include/model_skinned.sh>
-#include <../../../include/model_position.sh>
+#include <../../../include/model_transforms.sh>
 
 void main() {
-	gl_Position = getTranslatedPos(boneTransform(a_indices, a_weight, a_position), a_texcoord0);
+	gl_Position = applyPosTransforms(boneTransform(a_indices, a_weight, a_position), a_texcoord0);
 
 	v_texcoord0 = a_texcoord0;
 	v_color0 = a_color0;
