+++ render/shaders/model/unlit/fs_model_unlit.sc
@@ -1,6 +1,7 @@
 $input v_color0, v_texcoord0
 
 #include <bgfx_shader.sh>
+#include <../../include/model_transforms.sh>
 
 SAMPLER2D(s_texColor, 0);
