+++ render/shaders/model/unlit/vs_model_unlit.sc
@@ -2,10 +2,10 @@ $input a_position, a_color0, a_texcoord0
 $output v_color0, v_texcoord0
 
 #include <bgfx_shader.sh>
-#include <../../include/model_position.sh>
+#include <../../include/model_transforms.sh>
 
 void main() {
-    gl_Position = getTranslatedPos(a_position, a_texcoord0);
+    gl_Position = applyPosTransforms(a_position, a_texcoord0);
 
     v_color0 = a_color0;
     v_texcoord0 = a_texcoord0;
