+++ render/src/gizmos.cpp
@@ -1,5 +1,6 @@
 
 #include <rawrbox/render/gizmos.hpp>
+#include <rawrbox/render/model/utils/mesh.hpp>
 #include <rawrbox/utils/string.hpp>
 
 namespace rawrbox {
@@ -12,7 +13,7 @@ namespace rawrbox {
 	uint32_t GIZMOS::addGizmo(const std::string& type, const rawrbox::Vector3f& pos) {
 		auto id = ++_ID;
 
-		auto mesh = _gizmos->generatePlane({}, {0.25F, 0.25F});
+		auto mesh = rawrbox::MeshUtils<>::generatePlane({}, {0.25F, 0.25F});
 		mesh.setPos(pos);
 		mesh.setName(fmt::format("GIZMO-{}", ++_ID));
